requirements.txt
asgiref==3.4.1
Django==4.0
django-cors-headers==3.10.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
